# api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Global News Dashboard API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/admin/diagnostics/universe:
    get:
      summary: Get universe diagnostics (ADMIN)
      security:
        - ApiKeyAuth: []
  /api/admin/diagnostics/universe/rebuild:
    post:
      summary: Trigger universe rebuild now (ADMIN)
      security:
        - ApiKeyAuth: []
  /api/admin/diagnostics/ticker-alias:
    get:
      summary: Get ticker alias diagnostics (ADMIN)
      security:
        - ApiKeyAuth: []
  /api/admin/diagnostics/ticker-alias/verify:
    post:
      summary: Trigger ticker alias verification now (ADMIN)
      security:
        - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY